Rename `onthology` -> `ontology`
